Patch 3017253 by Sébastien Le Callonne
Two links in the “Toolbox” box on the left-hand side appear are set in different font 
to the other links in the same box. This is because their font-family is set to 
"Arial, Helvetica" whereas the other inherit the default font (specified as sans-serif).
The patch removes "Arial, Helvetica" so that these 2 links get the font-family 
"sans-serif" like the others.